Refactor Nostradamus
Moving logic of convert to nostradamus class
