+++ lib/nostradamus/version.rb
@@ -1,4 +1,4 @@
-module Nostradamus
+class Nostradamus
   module Version
     STRING = "0.0.3"
   end
